requirements.txt
Django==1.9.4
django-allauth==0.25.2
django-filter==0.13.0
django-rest-auth==0.7.0
django-rest-swagger==0.3.5
djangorestframework==3.3.2
Markdown==2.6.5
MySQL-python==1.2.5
mysqlclient==1.3.7
oauthlib==1.0.3
Pillow==3.1.1
python-openid==2.2.5
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
uWSGI==2.0.12
